Move to circuit_library, add in classes.rst
